some code changes

+++ src/components/Cart/Cart.js
@@ -1,44 +1,43 @@
-import { deleteDoc, getDocs, doc, getDoc } from 'firebase/firestore';
+import { deleteDoc, doc } from 'firebase/firestore';
 import { db } from '../../firebase/config';
 import React, { useEffect, useState } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { useNavigate, useParams } from 'react-router-dom';
-// import { connect } from 'react-redux';
+import { useNavigate } from 'react-router-dom';
+import { useUserAuth } from '../../context/UserAuthContext';
 
 const Cart = () => {
     const [cart, setCart] = useState([]);
     const navigate = useNavigate();
-    // const { id } = useParams();
-    const items = useSelector(state => {
-        return state.cartItems
-    });
     const dispatch = useDispatch();
-    // const itemQuantity = useSelector(state => state.itemQty);
+    const { loggedUser } = useUserAuth();
+    const items = useSelector(state => state.cartItems);
     const total = useSelector(state => state.totalPrice);
 
-    const removeFromCart = (itemID,itm) => {
-        console.log("item id = ", itemID);
-        const cartID=itm.cartId
-        // console.log("cart id = ",cartID);
-        const itmPrice=itm.qty*itm.price;
-        dispatch({
-            type: 'removeFromCart',
-            payload: {
-                id: itemID,
-                netAmount:itmPrice
-            }
-        });
-        deleteDoc(doc(db, "cart", cartID)).then(() => {
-            alert("Item deleted");
-        })
-        // console.log("deleted", itemID);
+    const removeFromCart = (itemID, itm) => {
+        const cartID = itm.cartId;
+        const itmPrice = itm.qty * itm.price;
+        if (loggedUser.uid === itm.userID) {
+            dispatch({
+                type: 'removeFromCart',
+                payload: {
+                    id: itemID,
+                    netAmount: itmPrice
+                }
+            });
+            deleteDoc(doc(db, "cart", cartID)).then(() => {
+                alert("Item deleted");
+            })
+        }
+        else {
+            alert("You are not authorised to perform this")
+        }
     }
 
     useEffect(() => {
         if (items.length === 0) {
             navigate("/cart/empty")
         }
-        
+
         else {
             setCart(items)
         }
@@ -92,7 +91,7 @@ const Cart = () => {
                                             <div className="card-body">
                                                 <button
                                                     className="btn btn-danger"
-                                                    onClick={() => removeFromCart(item.id,item)}>
+                                                    onClick={() => removeFromCart(item.id, item)}>
                                                     Remove From Cart
                                                 </button>
                                             </div>
